refactor surveyScreen
extract list of categories into own component
